Navbar fix and addition of user servises
