+++ NOTEA/Services/IUserService.cs
@@ -1,0 +1,10 @@
+﻿using NOTEA.Models;
+
+namespace NOTEA.Services
+{
+    public interface IUserService
+    {
+        public void SaveUsers(UserListModel userList);
+        public UserListModel LoadUsers();
+    }
+}
