+++ NOTEA/Views/Conspect/ConspectList.cshtml
@@ -1,5 +1,4 @@
-
-﻿@inject IHttpContextAccessor context
+@inject IHttpContextAccessor context
 ﻿@using Newtonsoft.Json;
 @model NOTEA.Models.ConspectListModel<ConspectModel>
