+++ NOTEA/Views/Shared/_Layout.cshtml
@@ -38,8 +38,8 @@
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="Conspect" asp-action="UploadConspect">Upload conspects</a>
                         </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="User" asp-action="LogIn">Log out</a>
+                        <li class="nav-item ms-auto">
+                           <a class="nav-link text-dark" asp-area="" asp-controller="User" asp-action="LogIn">Log out</a>
                         </li>
                     </ul>
                 </div>
